[FE] add update function
