+++ frontend/src/Requests/index.ts
@@ -17,6 +17,12 @@ export const removeMotorcycleCard = async (id: string) => {
   return response;
 };
 
+export const getMotorcycleCardById = async (id: string) => {
+  const response = await axios.get(`${API_V1_URL}/motorcycleCard/${id}`);
+
+  return response;
+};
+
 export const getCatalogMotorcycle = async () => {
   const response = await axios.get(`${API_V1_URL}/motorcycleCards`);
 
@@ -30,12 +36,21 @@ export const getMotorcycleCard = async (id: string) => {
 };
 
 export const uploadMotorcycleCardData = async (data: FormData) => {
-  const response = await fetch(`${API_V1_URL}/motorcycleCards`, {
-    method: 'POST',
-    body: data,
-  });
-  const jsonResponse = await response.json();
-  return jsonResponse;
+  const response = await $api.post('motorcycleCards', data);
+
+  return response;
+};
+
+export const updateMotorcycleCard = async ({
+  data,
+  id,
+}: {
+  data: FormData;
+  id: string;
+}) => {
+  const response = await $api.put(`/motorcycleCard/${id}`, data);
+
+  return response;
 };
 
 export const createUserProfile = async (data: ICreateUserProfileData) => {
@@ -57,8 +72,6 @@ export const loginRequest = async (data: ILoginForm) => {
 };
 
 export const activationRequest = async (activationLink: string) => {
-  //TODO: do need add $api?
-  // const response = await $api.get(`/activation/${activationLink}`);
   const response = await axios.get(
     `${API_V1_URL}/activation/${activationLink}`,
   );
