+++ frontend/src/UrlsConfig.ts
@@ -4,8 +4,7 @@ export const AboutUsUrl = 'about-us';
 export const ContactsUrl = 'contacts';
 export const CatalogMotorcyclesUrl = 'catalog-motorcycles';
 export const CreateMotorcycleCartUrl = 'create-motorcycle';
-export const AdminPageUrl = 'admin-page';
-// TODO: understand line bellow, why we don't us /:id
+export const AdminPageUrl = '/admin-page';
 export const UpdateMotorcycleCardUrl = '/update-motorcycle';
 export const UpdateMotorcycleCardIdUrl = '/update-motorcycle/:id';
 export const MotorcycleInfoUrl = 'motorcycle';
