+++ frontend/src/assets/scss/base/_variables.scss
@@ -68,3 +68,7 @@ $pc-screen: 720px;
 .white-background {
   background-color: $white;
 }
+
+.requestError {
+  color: $red-extra;
+}
