+++ frontend/src/components/FormCreateMotorcycleCard/style.scss
@@ -1,0 +1,250 @@
+@import '../../assets/scss/base/variables';
+
+.motorcycleCard__form {
+  .form__input-input,
+  .form__select-select {
+    width: 100%;
+    padding: 0 10px;
+    border-radius: 4px;
+    height: 40px;
+    margin-bottom: 10px;
+    background-color: $grey-100;
+    border: 1px $grey-400 solid;
+    font-size: 18px;
+  }
+
+  .form__select {
+    &-select {
+      cursor: pointer;
+      font-size: 16px;
+    }
+
+    &-option {
+      font-size: 16px;
+    }
+  }
+
+  .form__checkbox {
+    &-block {
+      display: flex;
+      justify-content: space-around;
+      height: 28px;
+      margin-bottom: 30px;
+
+      .checkboxBlock {
+        width: 45px;
+        height: 45px;
+        border: 1px $black solid;
+        border-radius: 4px;
+        padding-top: 7px;
+        padding-left: 7px;
+      }
+
+      .checkboxItem {
+        position: absolute;
+        z-index: -1;
+        opacity: 0;
+      }
+
+      .checkboxItem + .checkboxLabel {
+        cursor: pointer;
+        display: inline-flex;
+        align-items: center;
+        -webkit-user-select: none;
+        -moz-user-select: none;
+        -ms-user-select: none;
+        user-select: none;
+      }
+
+      .checkboxItem + .checkboxLabel::before {
+        content: '';
+        display: inline-block;
+        width: 28px;
+        height: 28px;
+
+        border: 1px solid #c3c3c3;
+      }
+
+      .checkboxItem:checked + .checkboxLabel::before {
+        width: 40px;
+        height: 40px;
+        margin: -2px;
+      }
+
+      .checkboxItem:not(:disabled):not(:checked)
+        + .checkboxLabel:hover::before {
+        border-color: rgba(0, 0, 255, 0.33);
+      }
+
+      .checkboxItem:focus + .checkboxLabel::before {
+        box-shadow: 0 0 0 0.2rem rgba(0, 0, 255, 0.125);
+      }
+      .checkboxItem:focus:not(:checked) + .checkboxLabel::before {
+        border-color: #c3c3c3;
+      }
+      .checkboxItem:disabled + .checkboxLabel::before {
+        background-color: black;
+      }
+    }
+  }
+
+  .dndBlock {
+    display: block;
+
+    &-title {
+      font-size: 18px;
+      margin-bottom: 8px;
+    }
+
+    &-paragraph {
+      color: $grey-600;
+      margin-bottom: 8px;
+    }
+
+    &-inputDnd {
+      display: block;
+      padding: 10px;
+      cursor: pointer;
+      margin-bottom: 10px;
+    }
+
+    &-preview {
+      margin: 18px 0 14px;
+      display: flex;
+      flex-wrap: nowrap;
+      justify-content: space-between;
+      max-height: 260px;
+
+      .preview-img {
+        width: 50%;
+      }
+
+      .preview-info {
+        width: 50%;
+        display: flex;
+        flex-wrap: wrap;
+        justify-content: right;
+
+        &-button {
+          width: 120px;
+          height: 40px;
+          border: 2px $grey-600 solid;
+          background-color: $grey-200;
+          transform: skew(-10deg);
+          cursor: pointer;
+        }
+
+        &-button:hover {
+          background-color: $grey-300;
+        }
+      }
+      .preview-info > p {
+        display: block;
+        width: 100%;
+        margin: 10px 0;
+        text-align: right;
+      }
+    }
+  }
+
+  .release {
+    border: 3px solid $grey-400;
+    background-image: url('../../assets/img/pattern.png');
+    min-height: 154px;
+  }
+
+  .attach {
+    border: 3px dashed $grey-400;
+    background-image: url('../../assets/img/pattern.png');
+  }
+
+  .dnd-inner {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    align-items: center;
+    padding-top: 10px;
+
+    .dnd-uploadSvg {
+      width: 40px;
+      height: 40px;
+      fill: $deep-blue;
+      display: block;
+      margin: 10px auto;
+    }
+
+    .inner-title {
+      font-size: 14px;
+      color: $black;
+      text-align: center;
+    }
+
+    .inner-paragraph {
+      font-size: 14px;
+      color: $grey-600;
+      text-align: center;
+    }
+
+    .paragraph-release {
+      display: block;
+      font-size: 18px;
+      font-weight: 700;
+      height: 94px;
+      padding-top: 20px;
+    }
+
+    .dnd-input {
+      width: 100%;
+      margin: 0 auto;
+      opacity: 0;
+      z-index: -1;
+    }
+  }
+
+  .buttonBlock {
+    display: flex;
+    justify-content: space-evenly;
+    padding-top: 20px;
+
+    .buttonBlock-submit,
+    .buttonBlock-cancel {
+      min-width: 200px;
+      transform: skew(-10deg);
+      text-transform: uppercase;
+      font-family: 'Ubuntu', cursive;
+      background-color: rgb(173, 3, 3);
+      color: white;
+      font-weight: 400;
+      font-size: 20px;
+      letter-spacing: 3px;
+      height: 40px;
+      padding: 4px 6px;
+      cursor: pointer;
+      border: none;
+    }
+
+    .buttonBlock-submit:hover {
+      background-color: rgb(141, 4, 4);
+    }
+
+    .buttonBlock-cancel:hover {
+      background-color: $grey-200;
+    }
+
+    .buttonBlock-cancel {
+      background-color: $white;
+      border: 2px $black solid;
+      color: $black;
+    }
+  }
+}
+
+.form__error {
+  color: $red-extra;
+  margin-bottom: 6px;
+}
+
+.form__label {
+  font-size: 18px;
+  margin-bottom: 6px;
+}
