+++ frontend/src/pages/AdminPage/style.scss
@@ -3,7 +3,7 @@
 .wrapper {
   background-color: #f3f0e7;
   padding: 20px 0;
-  height: 100vh;
+  height: calc(100vh - 60px);
 }
 
 .adminPage__table {
