+++ frontend/src/pages/CatalogMotorcycles/index.tsx
@@ -2,7 +2,7 @@ import { useQuery } from 'react-query';
 
 import { IMotorcycleCard } from './types';
 import MotorcycleCard from '../../components/MotorcycleCard';
-import { getCatalogMotorcycle } from '../../Requests';
+import { getCatalogMotorcycle } from '../../requests';
 
 import './style.scss';
