+++ frontend/src/pages/Registration/index.tsx
@@ -16,8 +16,8 @@ import {
   IRegistrationForm,
   ISuccessCreateUserProfile,
 } from './types';
-import { createUserProfile } from '../../Requests';
-import { IServerError } from '../../Requests/types';
+import { createUserProfile } from '../../requests';
+import { IServerError } from '../../requests/types';
 
 import './style.scss';
