+++ frontend/src/pages/UpdateMotorcycleCard/style.scss
@@ -1,0 +1,1 @@
+@import '../../assets/scss/base/variables';
