+++ frontend/src/pages/UpdateMotorcycleCard/types.ts
@@ -1,0 +1,4 @@
+import { INewMotorcycleCard } from '../../components/FormCreateMotorcycleCard/types';
+
+export interface IUpdateMotorcycleCard
+  extends Omit<INewMotorcycleCard, 'uploadImage'> {}
